dao id改为Integer类型

+++ shop-dao/src/main/java/com/masiis/shop/dao/platform/product/ComSpuMapper.java
@@ -27,7 +27,7 @@ public interface ComSpuMapper {
      * 添加一条记录
      * @param comSpu
      */
-    Long insert(ComSpu comSpu);
+    Integer insert(ComSpu comSpu);
 
     /**
      * 根据id更新一条记录
